Allow for multiple manual source files
MANSRC can now be set to a spece-separated list of manual source files.
In a similar way to SRCS in the Basic makefile, the first of these is
used as the source for passing to ManTools or XMLMan, and the rest are
used as dependencies for the Make rule.

Internally, MANFILE is set to the leaf of the first entry in MANSRC and
then used as the filename to use for the .XML tests when determining
whether to use XMLMan or ManTools for the build.